refactor: styles general refactor
