--- app/routes/_home.welcome._index/action.ts
@@ -1,12 +1,0 @@
-import { ActionFunctionArgs } from '@remix-run/node';
-
-export async function action({
-	request,
-}: ActionFunctionArgs): Promise<Response> {
-	const formData = await request.formData();
-	const username = formData.get('username');
-
-	console.log('FormData:', formData);
-
-	return Response.json({ message: `Hello, ${username}!` });
-}
