+++ app/shared/components/layout/app-footer/app-footer.module.css
@@ -1,3 +1,3 @@
 .appFooter {
-	@apply bg-gray-800 text-white text-center py-4;
+	@apply bg-gray-800 text-white text-center py-4 h-16;
 }
