+++ package.json
@@ -42,7 +42,7 @@
 		"@types/react": "18.2.20",
 		"@types/react-dom": "18.2.7",
 		"autoprefixer": "10.4.19",
-		"typescript": "5.1.6",
+		"typescript": "5.8.2",
 		"vite": "5.1.0",
 		"vite-tsconfig-paths": "4.2.1"
 	},
