+++ tailwind.config.js
@@ -3,10 +3,10 @@ const { heroui } = require('@heroui/theme');
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: [
-    "./app/**/*.{jsx,js,ts,tsx,css}",
-    "./node_modules/@heroui/theme/dist/components/(button|input|ripple|spinner|form).js"
-  ],
+	content: [
+		'./app/**/*.{jsx,js,ts,tsx,css}',
+		'./node_modules/@heroui/theme/dist/components/(button|input|ripple|spinner|form).js',
+	],
 	theme: {
 		extend: {},
 	},
