+++ tsconfig.json
@@ -8,7 +8,7 @@
 		"**/.client/**/*.tsx"
 	],
 	"compilerOptions": {
-		"lib": ["DOM", "DOM.Iterable", "ES2022"],
+		"lib": ["DOM", "DOM.Iterable", "ESNext"],
 		"types": ["@remix-run/node", "vite/client"],
 		"isolatedModules": true,
 		"esModuleInterop": true,
@@ -16,7 +16,7 @@
 		"module": "ESNext",
 		"moduleResolution": "Bundler",
 		"resolveJsonModule": true,
-		"target": "ES2022",
+		"target": "ESNext",
 		"strict": true,
 		"allowJs": true,
 		"skipLibCheck": true,
